Migrate base meals table
