+++ src/frontend/js/app.js
@@ -11,6 +11,4 @@ const App = () => (
   </Provider>
 );
 
-stores.security.checkLoggedIn();
-
 render(<App />, document.getElementById('container'));
