+++ src/frontend/js/components/DishPicker.js
@@ -1,0 +1,9 @@
+import { inject, observer } from 'mobx-react';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+export default class DishPicker extends React.Component {
+  render() {
+    return <p>Dish Picker</p>;
+  }
+}
